Moved some functions into a shared cog for Continuous Integration
init, dump_data, forget_data, generate_headers and get_token
